Changed the UX of the SessionCard (#8)
* Put the Date next to the Title of the "Video Card"

Date is now next to the Title, and I made the formatting of the Description "justify"
Also fixed typo

* Speaker is now sticky

* Changed date and description format

* Added a aspect ratio for the Player.

Didn't like the black bars on a large screen. It now scales accordingly

* Filters out the cards that do not have videoUrl

* Moved the videoUrl filtering to /archive/page.tsx

https://github.com/streamethorg/streameth-platform/pull/8#discussion_r1282978648

* Moved "Aspect-ratio" to <Player>

Firefox does work now :sunny:

Also fixed a issue where the scaling does not working if the screens height is super small

* Created a container for `<SessionsInfo>` and `<Player?`

* Prettier Files

* Added drop-shadow and changed scrolling

When you scroll the grid it would disappear out of thin air. It now hides behind the Filterbar, which has a drop shadow to give it more depth.

The scrolling of /archive already did this