🐛: Reverted changes on the library page and added state logic to clipping page (#772)
# Pull Request Info

## Description

I reverted a few changes that were breaking the library. It now creates
an `state` once the clip has been created. Showing it on the library
page while processing.

I also added a untested webhook if the video has failed to process. We
will have to wait until Livepeer fails to process once to examine the
webhook.

Fixes (issue)
#769 

## Type of change

Please delete options that are not relevant.

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published