🚸 Improved UX
If you were in page let's say 27 and then clicked on a tab item to
search for that topic, the query params would perform a `&` so most
likely there are no video of ZK (for example) in page 27.

The improvement consists on redoing the URL to it only matches the
topic, giving it priority over the page number.