⚡fixes #746 App takes you to login screen after being logged in for a while (#817)
# 😵 Post-Mortem 😵
Fixes  #746
## Summary
Fixed a bug that takes users to the login screen after being logged in
for a while.

## Impact
- **Services Affected**:Authentication service, user session management
- **User Impact**: Users will no longer be unexpectedly logged out after
an hour of inactivity. This change improves user experience by extending
the session duration to 30 days.

## Root Cause Analysis
The Privy token was set to expire after an hour, causing users to be
redirected to the login screen to log in again after the token expired.

## Resolution and Recovery
Instead of using the auth token, we now use the Privy refresh token. The
Privy refresh token takes 30 days to expire, can only be used once, and
is refreshed with a new one upon use. This change ensures that users
remain logged in for up to 30 days without interruption.

## Lessons Learned
1. Ensure token expiration times are aligned with user session
expectations to avoid unexpected logouts.
2. Implement a refresh token strategy to extend session durations
securely.
3. Regularly review and update authentication and session management
strategies to enhance user experience and security