🐛 Improve thumbnail upload; force 16:9 aspect ratio and fix overflow issues (#865)
## Description
To prevent overflow and future complications, this PR implements a
forced resize of images to 1280x720 pixels (16:9 ratio). If the original
image is smaller than this size, black bars are added to maintain the
aspect ratio. This approach aligns with YouTube's thumbnail guidelines.

## Type of change
- [x] Bug fix (non-breaking change which fixes an issue)

## Implementation Details
- Images are resized to 1280x720 pixels (16:9 ratio)
- Black bars are added if the original image of the thumbnail is smaller
/ bigger than 1280x720
- This change addresses overflow issues and potential future
complications
- Better error feedback for the user
- Give the `ImageUpload` component a `Dropzone`
- Made the size of the file smaller

## Future Work
Further in-depth work on this feature will be addressed in issue #868

## Reference
The approach taken is based on YouTube's guidelines:
- Make your thumbnail 1280 by 720 pixels (16:9 ratio)
- Ensure that your thumbnail is less than 2MB
- Use a JPG, PNG, or GIF file format